CHore: 파일이름 변경 및 스플리팅
signupForm 자체가 너무 무거워보여서 2가지로 분리하였으며 user를 member로 변경하였음
