+++ src/pages/signup/components/signupEssentailForm/styles.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const CheckBox = styled.div`
+  display: flex;
+  gap: 60px;
+  align-items: end;
+  justify-content: space-between;
+`;
